DACN/requirements.txt
gymnasium==0.29.1
gymnasium[classic-control]==0.29.1
numpy==1.26.4
pandas==2.2.2
scikit-learn
yfinance==0.2.44
matplotlib==3.9.2
plotly==5.24.1
stable-baselines3==2.3.0
sb3-contrib==2.3.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
optuna==3.6.1
tensorboard==2.17.0
pyyaml==6.0.2
rich==13.7.1
fastapi==0.110.1
uvicorn[standard]==0.30.1
streamlit==1.38.0
requests==2.32.3
websocket-client==1.8.0
python-dotenv==1.0.1
pytz==2024.1
openpyxl==3.1.5
google-genai
xgboost
